project/requirements.txt
fastapi==0.85.1
uvicorn==0.19.0
# new for postgres
asyncpg==0.27.0
# new for tortoise
tortoise-orm==0.19.2